requirements.txt
PyQt5
numpy
pyopengl
